requirements.txt
uWSGI==2.0.23
Flask==3.0.2
openai==1.66.0
python-dotenv==1.0.1
